person_django/myobject/requirements.txt
Django==1.11
olefile==0.44
Pillow==4.2.1
PyMySQL==0.7.11
pytz==2017.2
